icloud/requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fido2==2.0.0
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
jeepney==0.9.0
keyring==25.7.0
keyrings.alt==5.0.2
more-itertools==10.8.0
pycparser==2.23
PyExifTool==0.5.6
requests==2.32.5
SecretStorage==3.5.0
six==1.17.0
srp==1.0.22
tzlocal==5.3.1
urllib3==2.6.2
